# Propose `promextra.CounterVec` / `promextra.GaugeVec` to reduce allocations