--- a/pkg/util/promextra/vec.go
+++ b/pkg/util/promextra/vec.go
@@ -1,3 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
 package promextra
 
 import (