#### What this PR does

Each time we call WithLabelValues(...) we have to allocate a slice for that single argument that is immediately thrown away.

These allocations can't be seen on the profiling because it's the caller, not the `WithLabelValues` who performs an allocation, so they're spread across the entire codebase.

I propose to replace that in our hot paths by a wrapper that would reuse that slice through a pool. I replaced the usages in `pkg/ingester/metrics.go` as an example. Other usages would be done in follow up PRs. Most of our usages only need to pass one or two labels, so I did just that.

I didn't embed `prometheus.CounterVec` on purpose: this way it's easier to refactor without missing calls unchanged. I'm also leaving that field exported on purpose, as some usages may still need to access other methods.

I didn't want to spend too much time on naming, I'm open for discussion on that `promvec.Counter` could be another alternative.

#### Which issue(s) this PR fixes or relates to

None, just passing by.

#### Checklist

- [ ] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`