@colega 2023-10-02
Right, I guess I chose a bad example (I took a random file) for this. You're right, we should do that for userTSDB metrics where possible, as we could also make the lookup only once.

However, on the