@pstibrany 2023-10-03
Does it really help, or are you fighting the battle that compiler already optimizes? Where are the benchmarks?