@pstibrany 2023-10-03
> I'm not against this change

I am, unless it's provably and considerably faster. I don't think we need to make our codebase more complicated for miniscule improvements.