@colega 2023-10-03
My opinion is that this change isn't _more complicated_ as we're just wrapping with something that is more performant, but used in a similar way.

This can't be optimized by the compiler (at least at this point) as there's no way to know where that slice ends up.

Anyway, I hopelessly ran (because in the huge amount of work that Ingester.Push does, it's hard to find the impact of 1-2 less allocations) and I was surprised by the results. I even had to run it 50 times to make sure the result is consistent:

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
                │ /tmp/BenchmarkIngesterPush.old50 │  /tmp/BenchmarkIngesterPush.new50  │
                │              sec/op              │   sec/op     vs base               │
IngesterPush-16                        9.055µ ± 1%   8.982µ ± 1%  -0.81% (p=0.025 n=50)

                │ /tmp/BenchmarkIngesterPush.old50 │ /tmp/BenchmarkIngesterPush.new50 │
                │               B/op               │       B/op         vs base       │
IngesterPush-16                       1.064Ki ± 0%        1.033Ki ± 0%  -2.94% (n=50)

                │ /tmp/BenchmarkIngesterPush.old50 │ /tmp/BenchmarkIngesterPush.new50 │
                │            allocs/op             │     allocs/op      vs base       │
IngesterPush-16                         26.00 ± 0%          24.00 ± 0%  -7.69% (n=50)
```

I also ran `Benchmark_Ingester_PushOnError` getting similar results:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
                │ /tmp/Benchmark_Ingester_PushOnError.old │ /tmp/Benchmark_Ingester_PushOnError.new │
                │                 sec/op                  │        sec/op          vs base          │
IngesterPush-16                               9.046µ ± 2%             8.969µ ± 2%  ~ (p=0.579 n=10)

                │ /tmp/Benchmark_Ingester_PushOnError.old │ /tmp/Benchmark_Ingester_PushOnError.new │
                │                  B/op                   │       B/op        vs base               │
IngesterPush-16                              1.064Ki ± 0%       1.033Ki ± 0%  -2.94% (p=0.000 n=10)

                │ /tmp/Benchmark_Ingester_PushOnError.old │ /tmp/Benchmark_Ingester_PushOnError.new │
                │                allocs/op                │    allocs/op      vs base               │
IngesterPush-16                                26.00 ± 0%         24.00 ± 0%  -7.69% (p=0.000 n=10)
```