@colega 2023-10-03
> I think the canonical way to solve this problem is getting the WithLabelValues() reference once

I took a quick look at doing this:
- It's not an easy change because of all the userTSDB lifecycle. Can be done, but has to be made with extreme caution to avoid keeping references to deleted metrics.
- Some `WithLabelValues` calls would still stay: active trackers names, discarded samples groups, etc.