@colega 2023-10-04
@pstibrany pointed out that this should be optimized by the compiler, and @bboreham pointed to how it can be fixed, so this is superseded by the fix in https://github.com/prometheus/client_golang/pull/1360