@pstibrany 2023-10-05
> @bboreham pointed to how it can be fixed, so this is superseded by the fix in [prometheus/client_golang#1360](https://github.com/prometheus/client_golang/pull/1360)

That's quite interesting fix, very nice work!