@colega 2023-10-05
I tried benchmarking that fix vendored to Mimir and :rocket: :+1: 

```
goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/ingester
cpu: 11th Gen Intel(R) Core(TM) i7-11700K @ 3.60GHz
                │   old.txt   │              new.txt               │
                │   sec/op    │   sec/op     vs base               │
IngesterPush-16   7.508µ ± 2%   7.394µ ± 1%  -1.53% (p=0.000 n=10)

                │  old.txt   │              new.txt               │
                │    B/op    │    B/op     vs base                │
IngesterPush-16   960.5 ± 0%   831.0 ± 0%  -13.48% (p=0.000 n=10)

                │  old.txt   │              new.txt               │
                │ allocs/op  │ allocs/op   vs base                │
IngesterPush-16   18.00 ± 0%   10.00 ± 0%  -44.44% (p=0.000 n=10)
```